# Conflicts:
#	src/main/java/com/project/wekiosk/category/Controller/CategoryController.java
#	src/main/java/com/project/wekiosk/category/domain/Category.java
#	src/main/java/com/project/wekiosk/category/dto/CategoryDTO.java
#	src/main/java/com/project/wekiosk/category/repository/CategoryRepository.java
#	src/main/java/com/project/wekiosk/category/service/CategoryService.java
#	src/main/java/com/project/wekiosk/category/service/CategoryServiceImpl.java
#	src/main/java/com/project/wekiosk/option/domain/Options.java
#	src/main/java/com/project/wekiosk/order/controller/OrdersController.java
#	src/main/java/com/project/wekiosk/product/controller/ProductController.java
#	src/main/java/com/project/wekiosk/product/domain/Product1.java
#	src/main/java/com/project/wekiosk/product/domain/ProductImage.java
#	src/main/java/com/project/wekiosk/product/dto/ProductDTO.java
#	src/main/java/com/project/wekiosk/product/repository/ProductRepository.java
#	src/main/java/com/project/wekiosk/product/search/ProductSearchImpl.java
#	src/main/java/com/project/wekiosk/product/service/ProductService.java
#	src/main/java/com/project/wekiosk/product/service/ProductServiceImpl.java
#	src/main/resources/application.properties

+++ src/main/java/com/project/wekiosk/category/Controller/CategoryController.java
@@ -37,12 +37,18 @@ public class CategoryController {
         return new ResponseEntity<>(category, HttpStatus.OK);
     }
     @PutMapping("/{cateno}/modify")
-    public ResponseEntity<Void> updateCategory(@PathVariable Long cateno, @RequestBody CategoryDTO categoryDTO) {
+    public ResponseEntity<String> updateCategory(@PathVariable Long cateno, @RequestBody CategoryDTO categoryDTO) {
         categoryService.updateCategory(cateno, categoryDTO);
         return new ResponseEntity<>(HttpStatus.OK);
     }
     @DeleteMapping("/{cateno}/delete")
-    public ResponseEntity<Void> deleteCategory(@PathVariable Long cateno) {
+    public ResponseEntity<String> deleteCategory(@PathVariable Long cateno) {
+        boolean hasProducts = categoryService.hasProductsInCategory(cateno);
+
+        if (hasProducts) {
+            return ResponseEntity.badRequest().body("해당 카테고리에 연결된 상품이 있어 삭제할 수 없습니다.");
+        }
+
         categoryService.deleteCategory(cateno);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
